chapter_003/src/main/java/ru/job4j/collection/UniqueText.java
package ru.job4j.collection;

import java.util.Arrays;
import java.util.HashSet;

/**
 * Created by Intellij IDEA.
 * User: Vitaly Zubov.
 * Email: Zubov.VP@yandex.ru.
 * Version: $Id$.
 * Date: 24.12.2020.
 */
public class UniqueText {
    public static boolean isEquals(String originText, String duplicateText) {
        boolean rsl = true;
        String[] origin = originText.split(" ");
        String[] text = duplicateText.split(" ");
        HashSet<String> check = new HashSet<>(Arrays.asList(origin));
        for (String second : text) {
            if (!check.contains(second)) {
                rsl = false;
                break;
            }

        }
        return rsl;
    }
}
